% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_wwarn.R
\name{add_a_row_k13_clean}
\alias{add_a_row_k13_clean}
\title{Add a row to dataframe with all the same info except x and mut}
\usage{
add_a_row_k13_clean(df, x_new, mut_new)
}
\arguments{
\item{df}{WWARN dataframe that you want to add a row to.
Note: this will probably be a specific survey}

\item{x_new}{The value of x in the new row. For imputing wild type, x_new == n}

\item{mut_new}{The value of mut in the new row. For imputing wild type, these are the
WT encodings for codons not detected in the survey}
}
\description{
Add a row to dataframe with all the same info except x and mut
}
\details{
This function adds additional rows to the dataframe. It keeps all column variables
the same except for x and mut. This is a version of the function to be used in cleaning scripts
There is a different version of this used in the imputation hence the different names
TODO: get these aligned but most important for now that this works
}
