% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{add_tags_diff_studyID}
\alias{add_tags_diff_studyID}
\title{Tag rows in a data.frame for duplicates across different study_IDs}
\usage{
add_tags_diff_studyID(df)
}
\arguments{
\item{df}{A \code{data.frame} containing the data to be processed. It should include the following columns:
\itemize{
\item \code{study_ID}: Unique identifiers for each study.
\item \code{database}: The source database for each study (e.g., "WHO", "Pf7k", "GEOFF").
\item \code{publication_year}: The publication year of each study (used in Case 2).
}}
}
\value{
A \code{data.frame} with an added \code{keep_row} column, which contains the values "keep" or "remove".
The column indicates whether the corresponding row should be kept or removed based on the criteria for each case.
}
\description{
This function tags rows in a data.frame with a "keep_row" column to indicate which
records should be kept or removed based on duplicate study IDs across different databases.
The function handles four distinct cases:
}
\details{
\itemize{
\item \strong{Case 1:} All records are from the "WHO" or "Pf7k" databases:
\itemize{
\item Keep the first study_ID alphabetically and remove the rest.
}
\item \strong{Case 2:} All records are from the "GEOFF" database:
\itemize{
\item If duplicates exist between study_IDs starting with "S00" and "S01", keep the study_ID starting with "S01".
\item If all study_IDs start with "S00" or "S01", keep the most recent publication based on the \code{publication_year}.
}
\item \strong{Case 3:} Mixed records from "GEOFF" and either "WHO" or "Pf7k":
\itemize{
\item Keep records from "GEOFF" and remove records from "WHO" or "Pf7k".
}
\item \strong{Case 4:} Mixed records from "Pf7k" and "WHO" only:
\itemize{
\item Keep records from "Pf7k" and remove records from "WHO".
}
}
}
