% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{add_tags_same_studyID}
\alias{add_tags_same_studyID}
\title{Add tag to rows in data.frame for duplicates within the same studyID}
\usage{
add_tags_same_studyID(df)
}
\arguments{
\item{df}{A data.frame where rows should be tagged as "keep" or "remove".}
}
\value{
A data.frame with the \code{keep_row} column indicating if the study should be kept or removed.
Returns \code{NULL} if the rounded coordinates for GEOFF studies are not the same.
}
\description{
This function tags rows for removal based on duplicate geographic coordinates
within the same study ID. If the geographic coordinates (longitude and latitude)
are the same when rounded to the first decimal place for GEOFF studies, it keeps the
first row and removes the others. If the coordinates are not the same, it returns \code{NULL}.
For all other studies, the first row is kept, and the others are removed.
}
