% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_invalid_date.R
\name{adjust_invalid_date}
\alias{adjust_invalid_date}
\title{Adjust Invalid Dates}
\usage{
adjust_invalid_date(date_str, is_start = TRUE)
}
\arguments{
\item{date_str}{A string representing a date, which could be in one of the following formats:
"YYYY", "YYYY-MM", or "YYYY-MM-DD".}

\item{is_start}{A logical value. If \code{TRUE}, incomplete dates will be adjusted to the
start of the period (e.g., "YYYY" becomes "YYYY-01-01"). If \code{FALSE}, they will
be adjusted to the end of the period (e.g., "YYYY" becomes "YYYY-12-31").}
}
\value{
A valid \code{Date} object. If the date is invalid or unrecognized, it returns \code{NA}.
}
\description{
This function adjusts date strings that may be incomplete or invalid by
converting them to valid \code{Date} objects. It supports year-only, year-month,
and full year-month-day formats. For incomplete dates, it returns the first
or last day of the period, depending on whether the date is a start or end date.
}
\examples{
adjust_invalid_date("2021", is_start = TRUE)
# Returns: "2021-01-01"

adjust_invalid_date("2021-05", is_start = FALSE)
# Returns: "2021-05-31"

adjust_invalid_date("2021-05-15")
# Returns: "2021-05-15"

}
