% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_values_in_column.R
\name{check_values_in_column}
\alias{check_values_in_column}
\title{Check whether columns in the dataframe satisfy allowed values}
\usage{
check_values_in_column(data, column_name, allowed_values)
}
\arguments{
\item{data}{The preliminarily cleaned data where the format has been standardised, "nas" fixed etc.}

\item{column_name}{A string in " " with the name of the column of interest}

\item{allowed_values}{A vector of the values allowed within column_name}
}
\value{
either a warning or a message detailing whether the test passes or fails
}
\description{
This function looks at all values in a column and returns a warning if it is not within the set of allowed values
This function is generalisable to any column in the dataset and a defined suitable list
}
\examples{

data <- data.frame(country = c("Angola", "Burundi", "Eritrea", "Vietnam"))
countries <- data.frame(country = countrycode::codelist$country.name.en,
  continent = countrycode::codelist$continent)
african_countries <- countries \%>\%
  dplyr::filter(continent == "Africa") \%>\%
  dplyr::pull(country) \%>\% unique()
check_values_in_column(data = data, column_name = "country", allowed_values = african_countries)
pub <- data.frame(publication_status = c("peer_reviewed", "pre-print", "na"))
pub_status <- c("peer_reviewed", "preprint", NA)
check_values_in_column(data = pub, column_name = "publication_status", allowed_values = pub_status)
}
