% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_k13_range.R
\name{convert_k13_asterisk}
\alias{convert_k13_asterisk}
\title{Convert K13 Asterisk Mutation}
\usage{
convert_k13_asterisk(gene_mutation, mutation_key)
}
\arguments{
\item{gene_mutation}{A string representing the mutation in the format \code{"k13:codon:*"}.}

\item{mutation_key}{A dataframe containing mutation codon information with columns
\code{"PROTEIN"}, \code{"CODON"}, and \code{"REF"}.}
}
\value{
A string in the format \code{"K13:codon:amino_acid"}. If no matching codon is found, returns
\code{"K13:codon:NA"}.
}
\description{
This function converts a single-codon mutation string in the format \code{"k13:codon:*"}
into a standardized format by looking up the reference amino acid from a provided mutation key dataframe.
If the specified codon position does not exist in the mutation key, it will return \code{"K13:codon:NA"}.
}
\examples{
mutation_key <- data.frame(PROTEIN = "k13", CODON = 440:450, REF = letters[1:11])
convert_k13_asterisk("k13:442:*", mutation_key)
# Returns: "K13:442:c"

}
