% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_stave.R
\name{convert_stave}
\alias{convert_stave}
\title{Convert data into a stave compatible list of dataframes}
\usage{
convert_stave(data)
}
\arguments{
\item{data}{A wide dataset to be converted into a form compatible with \code{STAVE}}
}
\value{
A list of three dataframes easily convertible to be compatible with \code{STAVE} using stave$append_data
}
\description{
Convert data into a stave compatible list of dataframes
}
