% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_substudy_entries.R
\name{correct_substudy_entries}
\alias{correct_substudy_entries}
\title{Correct Known Bad Strings in the Substudy Column}
\usage{
correct_substudy_entries(substudy_column)
}
\arguments{
\item{substudy_column}{A character vector containing substudy entries that may
contain known typos or bad strings.}
}
\value{
A character vector with corrected substudy entries. Prints the number of
corrections made for each issue.
}
\description{
This function corrects known issues in the \code{substudy} column, such as converting
"day 0" to "day0extracted", "untreated_extracted" to "untreatedextracted",
and "untreadextracted" to "untreatedextracted". It counts and prints the number
of corrections made for each known issue.
}
\examples{
substudy_data <- c("day 0", "untreated_extracted", "untreadextracted", "day0extracted")
corrected_substudy <- scrub:::correct_substudy_entries(substudy_data)
# Output:
# Corrected 1 instances of day0extracted
# Corrected 1 instances of untreated_extracted
# Corrected 1 instances of untreadextracted


}
