% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_renamed_df.R
\name{create_renamed_df}
\alias{create_renamed_df}
\title{Create Dataframe with All Columns from the Original Data with Renamed Cols for WWARN rbind}
\usage{
create_renamed_df(df, mapping, default_database = "GEOFF")
}
\arguments{
\item{df}{A dataframe, typically the \code{master_table}, containing the original data.}

\item{mapping}{A named list that maps new column names to existing columns in the dataframe.
If the value is \code{NA}, the column will be filled with \code{NA}. If the column is "database", it will
be filled with the \code{default_database} value.}

\item{default_database}{A character string specifying the default value for the "database" column.
Defaults to "GEOFF".}
}
\value{
A dataframe that includes all original columns from the input dataframe as well as any additional
columns specified in the mapping.
}
\description{
This function creates a dataframe compatible with wwarn cleaned df rbind that includes all columns from the original
\code{master_table} and adds new columns as specified by the mapping. It preserves the original
columns and appends or replaces columns as needed.
}
\examples{
master_table <- data.frame(study_uid = c("s001", "s002"), gene_mutation = c("K13:580Y", "CRT:76T"))
mapping <- list(
  "gene" = "gene_mutation",
  "study" = "study_uid",
  "database" = NA
)
renamed_df <- scrub:::create_renamed_df(master_table, mapping)
print(renamed_df)


}
