% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{deduplicate}
\alias{deduplicate}
\title{Deduplicate Data in a Principled Way}
\usage{
deduplicate(df)
}
\arguments{
\item{df}{A data.frame to be deduplicated. It must contain the following columns:
\itemize{
\item \code{"iso3c"}: ISO3 country code
\item \code{"admin_0"}: Country-level administrative division
\item \code{"admin_1"}: First-level administrative division
\item \code{"site"}: Site name or identifier
\item \code{"lat"}: Latitude of the site
\item \code{"long"}: Longitude of the site
\item \code{"year"}: Year of the data entry
\item \code{"study_start_year"}: Starting year of the study
\item \code{"study_end_year"}: Ending year of the study
\item \code{"x"}: Numeric variable
\item \code{"n"}: Sample size
\item \code{"prev"}: Prevalence value
\item \code{"gene"}: Gene identifier
\item \code{"mut"}: Mutation identifier
\item \code{"gene_mut"}: Gene-mutation combination identifier
\item \code{"annotation"}: Annotation details
\item \code{"database"}: Source database
\item \code{"pmid"}: PubMed identifier for source
\item \code{"url"}: URL for additional source information
\item \code{"source"}: Source of the data
}}
}
\value{
A cleaned data.frame ready for use in STAVE
(see \url{https://github.com/mrc-ide/STAVE}).
}
\description{
This function deduplicates a dataset, ensuring each entry is unique based on
specified criteria, and returns a cleaned data frame suitable for input into STAVE.
}
