% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_wwarn.R
\name{extract_from_entry}
\alias{extract_from_entry}
\title{Impute the WT prevalence for non-mutant codons in WWARN dataset}
\usage{
extract_from_entry(entry)
}
\arguments{
\item{entry}{Character string. A string representing k13 SNP data in various formats.
This is programmed to handle double mutants, mixed infections etc.}
}
\value{
Integer. Processed mutation information into the codon position.
}
\description{
Impute the WT prevalence for non-mutant codons in WWARN dataset
}
\details{
This function takes the mut information and extracts the codon number.
This is applied over all of the values in the column to determine the markers found
in that study
}
\examples{
extract_from_entry("C580Y") == 580
extract_from_entry("C580Y_A481G") == c(580, 481)
extract_from_entry("C580Y/G") == 580
mut <- c("C580Y", "A481G", "R622I")
}
