% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{filter_duplicates_within_radius}
\alias{filter_duplicates_within_radius}
\title{Filter Dataframe for Duplicates Within a Specified Radius}
\usage{
filter_duplicates_within_radius(df, radius_km = 50)
}
\arguments{
\item{df}{A \code{data.frame} containing at least two columns:
\itemize{
\item \code{lon}: Longitude values of the geographic points.
\item \code{lat}: Latitude values of the geographic points.
}}

\item{radius_km}{A numeric value specifying the maximum allowed distance (in kilometers) between
all points. The default value is 50 km. This defines the "radius" within which all points must
fall in order to keep the dataframe.}
}
\value{
The original \code{data.frame} if all points are within the specified radius of each other.
If any point is outside the specified radius, the function returns \code{NULL}.
}
\description{
This function checks whether all geographic coordinates (longitude and latitude) in the given
dataframe are within a specified radius (in kilometers). If any point in the dataframe is outside
the radius, the function returns \code{NULL}, effectively excluding the dataframe. If all points are within
the specified radius, the original dataframe is returned unchanged.
}
