% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{group_signature_geoff}
\alias{group_signature_geoff}
\title{Generate a Group Signature for GEOFF Data}
\usage{
group_signature_geoff(df)
}
\arguments{
\item{df}{A data frame containing the following columns:
\itemize{
\item \code{variant_string}
\item \code{prev}
\item \code{total_num}
\item \code{variant_num}
\item \code{collection_year_start}
\item \code{collection_year_end}
}}
}
\value{
A single concatenated string signature for the group.
}
\description{
Creates a unique signature string for GEOFF datasets based on variant and prevalence data,
as well as sample sizes and collection years. This aids in identifying and distinguishing
unique groups in deduplication workflows.
}
