% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{group_signature_geoff_who_pf7k_wwarn}
\alias{group_signature_geoff_who_pf7k_wwarn}
\title{Generate a Group Signature for GEOFF vs WHO/Pf7k Comparison}
\usage{
group_signature_geoff_who_pf7k_wwarn(df)
}
\arguments{
\item{df}{A data frame containing the following columns:
\itemize{
\item \code{variant_string}
\item \code{collection_year_start}
\item \code{collection_year_end}
}}
}
\value{
A single string representing the group signature, with values concatenated
using underscores and records separated by \code{|}.
}
\description{
Creates a unique signature string to compare a GEOFF dataset to WHO or Pf7k datasets.
Uses either \code{collection_year_start} or \code{collection_year_end}, depending on which is
available. Useful in deduplication and data harmonization.
}
