% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_wwarn.R
\name{impute_study}
\alias{impute_study}
\title{Impute the WT in studies by applying impute_survey -- designed for cleaning WWARN}
\usage{
impute_study(study_df)
}
\arguments{
\item{study_df}{WWARN dataframe for this survey}
}
\value{
df Returns the df of a study with the imputed WT rows added
}
\description{
Impute the WT in studies by applying impute_survey -- designed for cleaning WWARN
}
\details{
This imputes the WT in the study.
Comprises of several parts:
\itemize{
\item Identifies the mutations in the study and extracts codon #
\item Finds min and max codons
\item Identifies validated markers in this range
\item Imputes each survey in the study and rbinds them together
}
}
