% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_wwarn.R
\name{impute_survey}
\alias{impute_survey}
\title{Impute the WT in surveys -- designed for cleaning WWARN}
\usage{
impute_survey(survey_df, impute_markers)
}
\arguments{
\item{survey_df}{WWARN dataframe for this survey (note: not the study)}

\item{impute_markers}{The df of markers to impute for the survey
When there are validated markers in the study, this is all of the WT markers
in the min-max codon range
In the case where only non-validated markers were found, this is the WT
study_markers == study_validated in {impute_study}}
}
\value{
df Returns the df of a survey with the imputed WT rows added
}
\description{
Impute the WT in surveys -- designed for cleaning WWARN
}
\details{
This imputes the WT in the survey.
}
