% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf7k_utils.R
\name{pf7k_format_k13_for_stave}
\alias{pf7k_format_k13_for_stave}
\title{Convert pf7k k13 SNP formats for Stave}
\usage{
pf7k_format_k13_for_stave(marker)
}
\arguments{
\item{marker}{Character. A string representing k13 SNP data in various formats.}
}
\value{
Character. Processed k13 SNP data in a standardised format for \code{STAVE}.
Returns \code{NA} if the input is missing.
}
\description{
Convert pf7k k13 SNP formats for Stave
}
\details{
This function processes k13 SNP data into a format compatible with the \code{STAVE} tool.
It handles different scenarios of input data, such as wild types, phased/unphased mutations,
and distinct haplotypes. The function accommodates variations in notation, including the use
of \verb{,}, \code{/}, and \code{*} in the input string.

NB FOR STAVE: "|" reflects phased

NB FOR STAVE "/" reflects unphased
\subsection{Input format conventions:}{
\itemize{
\item \strong{Uppercase letters}: Homozygous mutations.
\item \strong{Lowercase letters}: Heterozygous mutations.
\item \strong{Lowercase without a second mutation}: Heterozygous at the SNP level but same amino acid.
\item \verb{,}: Denotes separate, distinct haplotypes (different clones).
\item \code{/}: Denotes additional non-synonymous (NS) changes in the same clone.
\item \code{*}: Indicates unphased data for the sample.
}
}

\subsection{Output format:}{
\itemize{
\item \code{WT}: Wild type.
\item Mutations are prefixed with "k13-" and standardised.
\item Multiple haplotypes or NS changes are separated with \code{&&} for downstream splitting.
}
}
}
\examples{
pf7k_format_k13_for_stave("")  # Blank
pf7k_format_k13_for_stave("C580Y")  # Single SNP homozygous
pf7k_format_k13_for_stave("c580y")  # Single SNP heterozygous but same SNP
pf7k_format_k13_for_stave("C580Y,G449A")  # Distinct haplotypes
pf7k_format_k13_for_stave("C580Y/G449A")  # NS changes in same clone
pf7k_format_k13_for_stave("C580Y,G449A/I543T")  # Mixed haplotypes and NS changes
}
