% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_survey.R
\name{process_survey}
\alias{process_survey}
\title{Process a single survey ID to impute missing k13 mutation records}
\usage{
process_survey(each_survey)
}
\arguments{
\item{each_survey}{A character string representing a unique survey ID to be processed.}
}
\value{
A tibble containing imputed records for missing k13 mutations, or NULL if no
imputation is needed or the survey has missing/inconsistent k13 range data.
}
\description{
This function processes an individual survey ID from the master_table_formatted
dataset to identify and impute missing k13 mutation records. It checks the
coverage range from k13_min:k13_max in study_overview, extracts existing k13
mutations, and adds missing variants based  on the reference mutation key
defining variant_num and total_num from average of total_num of all k13 alleles
in a survey between the specified k13 range from the study overview.
}
\examples{
# Example usage (not run):
# process_survey("S0001")

}
