% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRuB.R
\name{SCRuB}
\alias{SCRuB}
\title{Decontaminate a set of samples
SCRuB removes contamination from an inputted set of samples using microbial source-tracking techniques.}
\usage{
SCRuB(
  data,
  metadata,
  control_order = NA,
  dist_threshold = 1.5,
  dist_metric = "euclidean",
  verbose = F
)
}
\arguments{
\item{data}{( n_samples + n_controls ) x n_taxa -- a count matrix representing the observed reads from the controls, and the samples to be decontaminated}

\item{metadata}{a metadata matrix of 2 columns (and a third optional column), where each row name is aligned to the `data` parameter row name. 
The first metadata column is boolean, denoting `TRUE` if the corresponding `data` row belongs to a sample representing a contamination source, and `FALSE` otherwise. 
The second column is a string, identifying the type of each sample, such that SCRuB can identifying which control samples should be grouped together. 
The third (and optional, but highly recommended) column is a string entry identifying the well location of the corresponding sample, which allows SCRuB to track well leakage.
This must be in a standard *LETTER**NUMBER* format, i.e. A3, B12, D4...}

\item{control_order}{vector, default NA. If specified, outlines the order of controls to run decontamination in. Input as a vector, of which each element must also be found in the metadata's second column. If not specified, all control types found in `metadata` will be run sequentially based on their order from that table.}

\item{dist_threshold}{float - Determines the maximum euclidean distance between samples and controls which SCRuB determines as potential sources of well leakage. This input is only used if the well location metadata is provided Default of 1.5}

\item{dist_metric}{string, default `euclidean`. The distance metric to be used when evaluating samples' physical distance. This input is used in the `stats` library's `dist` function; see their documentation for other options.}

\item{verbose}{boolean - if TRUE, SCRuB prints the log-likelihood of hte dataset thoughout each iteration.}
}
\value{
A list containing:
1) decontaminated_samples - a n_samples x n_taxa count matrix, representing the decontaminated samples
2) p - The fitted p parameter, as described in SCRuB's methods. 
An n_sample vector representing the estimate proportion of each observe sample that was not contamination
A dataset that had no contamination would have a p of 1s, while a dataset of entirely contamination would have a p of 0
3) inner_iterations -- results from SCRuB's intermediary steps, see the `Spatial_SCRUB` and `SCRUB_no_spatial` documentation for more information
}
\description{
Decontaminate a set of samples
SCRuB removes contamination from an inputted set of samples using microbial source-tracking techniques.
}
