% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{spatial_SCRUB}
\alias{spatial_SCRUB}
\title{Decontaminate a set of samples
SCRuB removes contamination from an inputted set of samples using microbial source-tracking techniques.}
\usage{
spatial_SCRUB(
  data,
  is_control,
  well_dists,
  dist_threshold = 1.5,
  print_loglikelihood = F,
  COVERAGE = 10000
)
}
\arguments{
\item{data}{( n_samples + n_controls ) x n_taxa -- a count matrix representing the observed reads from the controls, and the samples to be decontaminated}

\item{is_control}{boolean vecotr, length n_samples+n_controls -- a booolean vector indicating which rows of the data matrix correspond to negative controls, whose contents represent the contamination community to be removed from the non-control samples}

\item{well_dists}{numeric matrix. Deontes the spatial distance between each pairwise samples, names rows and columns must correst to the rownames of `data`}

\item{dist_threshold}{float - Determines the maximum euclidean distance between samples and controls which SCRuB determines as potential sources of well leakage. Default of 1.5}

\item{print_loglikelihood}{Boolean, TRUE of FALSE. Determines if SCRuB should print the calculated log-likelihood during each iteration}

\item{COVERAGE}{INT, Rarefaction depth to use in the parameter initializatin pipeline}
}
\value{
A list containing:
1) decontaminated_samples - a n_samples x n_taxa count matrix, representing the decontaminated samples
2) p - The fitted p parameter, as described in SCRuB's methods. 
An n_sample vector representing the estimate proportion of each observe sample that was not contamination
A dataset that had no contamination would have a p of 1s, while a dataset of entirely contamination would have a p of 0
3) alpha - The fitted alpha parameter, as described in SCRuB's methods. 
An n_control x ( n_sample + 1 ) matrix, representing the estimated contribution of the contaminant and each sample to each control, where the (n_sample + 1)th column represents the contribution from the contamination to the control.
Each row of alpha sums to 1, with each entry of the (n_sample + 1)th  column being 1 means there is zero estimated well leakge, while entries close to zero would indicate there is a high level of well leakage
4) gamma - the gamma parameter described in SCRuB's methods. An n_taxa vector representing the estimated relative abundance of the contamination community
5) loglikelihood - float. The log-likelihood of the inputted dataset based on SCRuB's fitted parameters.
}
\description{
Decontaminate a set of samples
SCRuB removes contamination from an inputted set of samples using microbial source-tracking techniques.
}
