% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partner_drug_utils.R
\name{standardise_mdr1_86}
\alias{standardise_mdr1_86}
\title{Standardise codon 86 of mdr1 gene}
\usage{
standardise_mdr1_86(x)
}
\arguments{
\item{x}{mdr1 mutation character string}
}
\value{
character string with standardises 86 codon
}
\description{
Standardise codon 86 of mdr1 gene
}
\details{
This function is designed to standardise the mdr1 mutations in codon 86 so that they can be
easily collapsed, combined, cleaned and the complement can be imputed
}
\examples{
# test this function
standardise_mdr1_86("mdr1:86:N") == "mdr1:86:N"
standardise_mdr1_86("mdr1:86:Y/N") == "mdr1:86:N/Y" 
standardise_mdr1_86("mdr1:86_184_1246:N_Y_D") == "mdr1:86:N"
standardise_mdr1_86("mdr1:86_184:Y/F_N") == "mdr1:86:Y" 

}
