% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{summarize_duplicates}
\alias{summarize_duplicates}
\title{Summarize Duplicates Across Multiple Dataframes}
\usage{
summarize_duplicates(duplicate_dfs, round_digits = 1)
}
\arguments{
\item{duplicate_dfs}{A list of dataframes, where each dataframe contains potential duplicate entries.
Each dataframe must have a \code{study_ID} column and optionally other relevant columns.}

\item{round_digits}{A numeric value indicating how many decimal places longitude and latitude values should be rounded to
before checking for duplicates. Default is 1.}
}
\value{
A dataframe summarizing duplicate pairs with the following columns:
\item{study_ID_1}{The study ID of the first entry in the duplicate dataframe.}
\item{study_ID_2}{The study ID of the second entry in the duplicate dataframe.}
\item{matching_columns}{A comma-separated list of columns that have identical values across duplicates.}
\item{variant_string, total_num, variant_num, prev, site_name, iso3c}{Columns that maintain their unique value if consistent across duplicates, otherwise marked as "not duplicate".}
\item{lon, lat}{Rounded longitude and latitude values if consistent across duplicates, otherwise marked as "not duplicate".}
\item{keep}{The study ID that should be retained based on the "keep_row" column. If no entry has "keep", it is set to NA.}
}
\description{
This function takes a list of dataframes containing potential duplicate records and summarizes their similarities and differences.
It identifies matching columns, checks for unique values in specific fields, rounds and compares latitude/longitude values,
and determines which study should be kept based on the "keep_row" column.
}
