% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwwarn_utils.R
\name{wwarn_format_k13_for_stave}
\alias{wwarn_format_k13_for_stave}
\title{Convert wwarn k13 SNP formats for Stave}
\usage{
wwarn_format_k13_for_stave(marker)
}
\arguments{
\item{marker}{Character. A string representing k13 SNP data in various formats.}
}
\value{
Character. Processed k13 SNP data in a standardised format for \code{STAVE}.
Returns \code{NA} if the input is missing or errors if unexpected type
}
\description{
Convert wwarn k13 SNP formats for Stave
}
\details{
This function processes k13 SNP data into a format compatible with the \code{STAVE} tool.

NB FOR STAVE: "|" reflects phased

NB FOR STAVE "/" reflects unphased
\subsection{Input format conventions:}{
\itemize{
\item \verb{_}: Denotes separate, distinct haplotypes (different clones).
\item \code{/}: Denotes additional non-synonymous (NS) changes in the same clone.
}
}

\subsection{Output format:}{
\itemize{
\item \code{WT}: Wild type.
\item Mutations are prefixed with "k13-" and standardised.
}
}
}
\examples{
wwarn_format_k13_for_stave("")  # Blank
wwarn_format_k13_for_stave("WT")  # WT
wwarn_format_k13_for_stave("P441L")  # Single SNP homozygous
wwarn_format_k13_for_stave("R528G/T")  # Mixed Infection
wwarn_format_k13_for_stave("T474I_V520I")  # Phased double mutant
}
