% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{add_tags_diff_studyID}
\alias{add_tags_diff_studyID}
\title{Add tag to rows in data.frame for duplicates between different studies}
\usage{
add_tags_diff_studyID(df)
}
\arguments{
\item{df}{A data.frame where rows should be tagged as "keep" or "remove".}
}
\value{
A data.frame with the keep_row column indicating if the study should
be kept or removed.
}
\description{
This function handles cases where duplicates exist across different study_IDs
by adding a column keep_row that determines which rows should be keep and
removed. There are 4 cases:
\itemize{
\item Case 1: All records are from WHO or Pf7k: keep the first study_ID
(alphabetically) and remove the others
\item Case 2: All records are from GEOFF: If duplicates are between study_IDs begin
with S00 and S01, pick S01 and if all study_IDs begin with S00 or S01,
pick most recent publication
\item Case 3: Mixture of GEOFF and WHO or Pf7k: Keep GEOFF data over WHO/Pf7k
\item Case 4: Mixture of Pf7k and WHO: keep Pf7k and remove WHO
}
}
