% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{handle_same_studyID_duplicates}
\alias{handle_same_studyID_duplicates}
\title{Add tag to rows in data.frame for duplicates within the same study}
\usage{
handle_same_studyID_duplicates(df)
}
\arguments{
\item{df}{A data.frame where rows should be tagged as "keep" or "remove".}
}
\value{
A data.frame with the keep_row column indicating if the study should
be kept or removed.
}
\description{
This function adds column keep_row and marks the first the first row in the group
to keep and remove the other rows which are duplicates.
}
